fix(notifications): include magic login link in join-approval email (Issue 882)
Thread the magic token minted on member promotion through
send_join_approval() into send_join_approval_email() so the approval
email carries a working login link, matching the magic_login email's
style. Applies to both manual full-approve and tentative-to-member
promotion via event check-in.

Remove the now-inaccurate "you'll receive a separate login link"
default body text since the link is now included directly.
