+++ backend/templates/emails/join_approval.html
@@ -4,7 +4,13 @@
   <div style="max-width: 480px; margin: 0 auto; background: #ffffff; border-radius: 8px; padding: 32px;">
     <p style="margin: 0 0 16px;">hi{% if display_name %} {{ display_name|lower }}{% endif %} —</p>
     <p style="margin: 0 0 16px;">you're fully approved — welcome to pda!</p>
-    <div style="margin: 0 0 8px; font-size: 13px; color: #666;">{{ message_body|linebreaks }}</div>
+    <div style="margin: 0 0 24px; font-size: 13px; color: #666;">{{ message_body|linebreaks }}</div>
+    <p style="margin: 0 0 24px;">
+      <a href="{{ magic_link_url }}" style="display: inline-block; background: #3c6939; color: #ffffff; text-decoration: none; padding: 12px 24px; border-radius: 6px; font-weight: 500;">log in to pda</a>
+    </p>
+    <p style="margin: 0 0 8px; font-size: 13px; color: #666;">or paste this into your browser:</p>
+    <p style="margin: 0 0 24px; font-size: 13px; word-break: break-all;"><a href="{{ magic_link_url }}" style="color: #3c6939;">{{ magic_link_url }}</a></p>
+    <p style="margin: 0; font-size: 13px; color: #666;">this link expires in 7 days.</p>
   </div>
   <p style="text-align: center; margin: 16px 0 0; font-size: 12px; color: #888;">— pda</p>
 </body>
