+++ backend/templates/emails/join_approval.txt
@@ -4,4 +4,10 @@ you're fully approved — welcome to pda!
 
 {% autoescape off %}{{ message_body }}{% endautoescape %}
 
+here's your one-time login link to set up your account:
+
+{{ magic_link_url }}
+
+this link expires in 7 days.
+
 — pda
